llm-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
accelerate==0.25.0
bitsandbytes==0.41.3
datasets==2.15.0
peft==0.7.1
trl==0.7.4
Pillow==10.0.1
numpy==1.24.3
psycopg2-binary==2.9.7
redis==5.0.1
gunicorn==21.2.0
werkzeug==2.3.7
openai==1.3.7
anthropic==0.7.8
ollama==0.1.7
huggingface-hub==0.19.4
tokenizers==0.15.0
sentencepiece==0.1.99
protobuf==4.25.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.33.1
moviepy==1.0.3
gtts==2.4.0
pydub==0.25.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
asyncio-mqtt==0.16.1
websockets==12.0
schedule==1.2.0
python-crontab==3.0.0
cryptography==41.0.8
jwt==1.3.1
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0
email-validator==2.1.0
python-slugify==8.0.1
unidecode==1.3.7
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langchain-experimental==0.0.47
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
wordcloud==1.9.2
tweepy==4.14.0
python-telegram-bot==20.7
discord.py==2.3.2
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.16.0
playwright==1.40.0
fake-useragent==1.4.0
proxyscrape==0.3.0
stem==1.8.2
pysocks==1.7.1
aiohttp==3.9.1
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.0
elasticsearch==8.11.0
kafka-python==2.0.2
pika==1.3.2
kombu==5.3.4
vine==5.1.0
click==8.1.7
rich==13.7.0
typer==0.9.0
fire==0.5.0
hydra-core==1.3.2
omegaconf==2.3.0
wandb==0.16.1
tensorboard==2.15.1
mlflow==2.8.1
dvc==3.32.0
great-expectations==0.18.5
evidently==0.4.11
whylogs==1.3.13
deepspeed==0.12.4
lightning==2.1.2
ray==2.8.1
optuna==3.4.0
hyperopt==0.2.7
scikit-optimize==0.9.0
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
statsmodels==0.14.0
scipy==1.11.4
networkx==3.2.1
igraph==0.11.3
plotly==5.17.0
dash==2.14.2
streamlit==1.28.2
gradio==4.8.0
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6
jupyterlab==4.0.9
voila==0.5.5
papermill==2.4.0
nbconvert==7.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
bandit==1.7.5
safety==2.3.5
pre-commit==3.6.0
